[content-sdk] Added catch handler
closes #54

This adds a simple catch handler. I have not used Ignition because I did not see any advantages, only disadvantages:

* Ignition is quite fat and installs bunyan/logging, which we don't need
* Ignition error format != JSON API format

The error format is not documented, but we don't want to break people's implementation, because the SDK is already 1.0.
We won't remove "err.request" and "err.config" for now and remove it in 2.0.